* labels now have a padding within them  * modified some upgrade descriptions  * personal shield now takes longer to regenerate

+++ src/main/java/de/sanandrew/mods/turretmod/api/client/tcu/ILabelElement.java
@@ -25,7 +25,7 @@ public interface ILabelElement
 
     default void doRenderTextured(ITurretInst turretInst, float maxWidth, float progress, FontRenderer stdFontRenderer) { }
 
-    default int getPriority() {
+    default short getPriority() {
         return 0;
     }
 }
